+++ libs/next/src/lib/components/dropdowns/select/select.options.ts
@@ -7,7 +7,7 @@ import { ZuiSelectIdentityMatcher, ZuiSelectSearchMatcher } from './select.model
 export interface ZuiSelectOptions<T> {
     readonly items: unknown[];
     readonly searchable: boolean;
-    readonly forceShowClearButton: boolean | null;
+    readonly forceClear: boolean | null;
     readonly label: string;
     readonly placeholder: string;
     readonly size: ZuiInputSize;
@@ -33,7 +33,7 @@ export const ZUI_SELECT_DEFAULT_OPTIONS: ZuiSelectOptions<unknown> = {
   items: [],
   searchable: false,
   outer: false,
-  forceShowClearButton: null,
+  forceClear: null,
   dropdownWidth: '100%',
   minDropdownHeight: 0,
   maxDropdownHeight: 342,